// generated/src/app/api-models/ar-tax-adjustment.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ArTaxAdjustment {
  seqAradjId: number;
  customerType: string;
  customerId: string;
  companyCode: string;
  transDate: Date;
  glRefCode: string;
  taxPct: number;
  taxAmt: number;
  statementStatus: string;
  invoiceNo: number;
  postDate: Date;
  seqGpbilId: number;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  taxType: string;
}